requirements.txt
ansicon==1.89.0
APScheduler==3.7.0
arrow==1.0.2
asgiref==3.2.7
beautifulsoup4==4.9.3
blessed==1.18.0
certifi==2020.4.5.1
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==3.0.6
django-cleanup==5.0.0
django-common-helpers==0.9.2
django-filter==2.3.0
django-hitcount==1.3.2
django-picklefield==3.0.1
django-rest-swagger==2.2.0
djangorestframework==3.11.0
djangorestframework-simplejwt==4.3.0
et-xmlfile==1.0.1
idna==2.9
itypes==1.2.0
jdcal==1.4.1
Jinja2==2.11.2
jinxed==1.0.1
kavenegar==1.1.2
MarkupSafe==1.1.1
numpy==1.20.1
openapi-codec==1.3.2
openpyxl==3.0.6
pandas==1.2.2
Pillow==8.0.0
PyJWT==1.7.1
python-dateutil==2.8.1
pytz==2020.1
requests==2.23.0
simplejson==3.17.0
six==1.15.0
soupsieve==2.2
sqlparse==0.3.1
typing-extensions==3.7.4.3
tzlocal==2.1
uritemplate==3.0.1
urllib3==1.25.9
wcwidth==0.2.5
xlrd==2.0.1
